ready for next stable release. Xuesong and Jinjin. 1) editing interface for signal movement data 2) data exchange interface for QEM data 3) real-time information user class ready 4) import openstreet map GPS 5) bug fixing for link moe

+++ trunk/Version1_01/NEXTA-GUI/Utility.cpp
@@ -65,7 +65,7 @@ void g_ProgramStop()
 };
 
 
-float g_GetPoint2Point_Distance(GDPoint p1, GDPoint p2)
+double g_GetPoint2Point_Distance(GDPoint p1, GDPoint p2)
 {
 return pow(((p1.x-p2.x)*(p1.x-p2.x) + (p1.y-p2.y)*(p1.y-p2.y)),0.5);
 }
